Added functionality to extract from multiple sources and removed preprocessing by adding those functionality to xsr_client.py
