+++ app/core/management/utils/xsr_client.py
@@ -1,32 +1,121 @@
-import json
+import io
 import logging
-import os
 
+import boto3
 import pandas as pd
-import requests
-from core.management.utils.preprocess import get_etca_bci, get_etca_bci_aetc
 
 logger = logging.getLogger('dict_config_logger')
 
 
-def extract_source():
-    """Extracting raw metadata from tables"""
-    logger.info('Extracting raw metadata from tables for further process')
+def get_aws_details():
+    """Get table data from the s3 bucket"""
 
-    ETCA_BCI_df = get_etca_bci()
-    ETCA_BCI_AETC_df = get_etca_bci_aetc()
-    return ETCA_BCI_df
+    # get object and file (key) from bucket
+    bucket = 'xsraetc'
+    file_name = 'AFCS_ETCA_Data.xlsx'
+    client = boto3.client('s3')
+    csv_obj = client.get_object(Bucket=bucket, Key=file_name)
+    table_data = csv_obj['Body'].read()
+    return table_data
+
+
+def get_list_table_names():
+    """Function to return table names for preprocessing extraction"""
+
+    table_list = ['ETCA_Course_Paragraph', 'ETCA_Course_Paragraph_711HPW',
+                  'ETCA_Course_Paragraph_ACC', 'ETCA_Course_Paragraph_AETC',
+                  'ETCA_Course_Paragraph_AETC2', 'ETCA_Course_Paragraph_AETC3',
+                  'ETCA_Course_Paragraph_AETC4', 'ETCA_Course_Paragraph_AFIT',
+                  'ETCA_Course_Paragraph_AFSOC', 'ETCA_Course_Paragraph_ANG',
+                  'ETCA_Course_Paragraph_AU', 'ETCA_Course_Paragraph_FT',
+                  'ETCA_Course_Paragraph_LT1K', 'ETCA_Course_Paragraph_Org_1K',
+                  'ETCA_Course_Paragraph_SAFIG',
+                  'ETCA_Course_Paragraph_USAFEC']
+
+    return table_list
+
+
+def get_etca_bci():
+    """ Extracting raw metadata from tables for further process"""
+
+    # Get table data from the s3 bucket
+    table_data = get_aws_details()
+
+    # Ingesting as dataframe values
+    etca_bci_df = pd.read_excel(io.BytesIO(table_data), sheet_name='ETCA_BCI',
+                                engine='openpyxl')
+
+    # Adding fields from tables to core table to enrich data
+    etca_bci_df = get_paragraph_data(etca_bci_df)
+
+    # Adding Resp_Org_ID to metadata
+    etca_bci_df['Resp_Org_ID'] = 'ETCA_BCI'
+
+    return etca_bci_df
+
+
+def get_etca_bci_aetc():
+    """ Extracting raw metadata from tables for further process"""
+
+    # Get table data from the s3 bucket
+    table_data = get_aws_details()
+
+    # Ingesting as dataframe values
+    etca_bci_aetc_df = pd.read_excel(io.BytesIO(table_data),
+                                     sheet_name='ETCA_BCI_AETC',
+                                     engine='openpyxl')
+
+    # Adding fields from tables to core table to enrich data
+    etca_bci_aetc_df = get_paragraph_data(etca_bci_aetc_df)
+
+    # Adding Resp_Org_ID to metadata
+    etca_bci_aetc_df['Resp_Org_ID'] = 'ETCA_BCI_AETC'
+
+    return etca_bci_aetc_df
+
+
+def get_paragraph_data(core_table_df):
+    """Adding fields from tables to enrich data in core tables"""
+
+    # Get table data from the s3 bucket
+    table_data = get_aws_details()
+
+    # Making a list of paragraph tables sheet names
+    table_list = get_list_table_names()
+    for table_name in table_list:
+        etca_course_paragraph_df = pd.read_excel(io.BytesIO(table_data),
+                                                 sheet_name=table_name,
+                                                 engine='openpyxl',
+                                                 usecols=['BCI_ID',
+                                                          'Paragraph_Heading',
+                                                          'Paragraph_Text'])
+        #  Retrieving rows from Dataframe with BCI_ID
+        etca_course_paragraph_df = etca_course_paragraph_df[
+            etca_course_paragraph_df["BCI_ID"].notnull()]
+
+        # Iterating over rows in Tables to find fields to add to the core table
+        for index, row in etca_course_paragraph_df.iterrows():
+            # Finding index for row in core table to be updated
+            index_val = core_table_df.index[core_table_df['BCI_ID'] ==
+                                            row['BCI_ID']]
+            # Adding updated to row with new field data
+            core_table_df.loc[core_table_df.index[index_val],
+                              row['Paragraph_Heading']] = row['Paragraph_Text']
+
+    return core_table_df
 
 
 def read_source_file():
     """Sending source data in dataframe format"""
-    logger.info("Retrieving data from XSR")
+    logger.info("Retrieving data from XSR for Extraction")
+
     # Function call to extract data from source repository
-    source_df = extract_source()
+    etca_bci_df = get_etca_bci()
+    etca_bci_aetc_df = get_etca_bci_aetc()
 
-    # # Changing null values to None for source dataframe
-    std_source_df = source_df.where(pd.notnull(source_df), None)
+    #  Creating list of dataframes of sources
+    source_list = [etca_bci_df, etca_bci_aetc_df]
 
-    # std_source_df.to_csv('etca_bci.csv')
-    # logger.debug("Sending source data in dataframe format for EVTVL")
-    return std_source_df
+    logger.debug("Sending source data in dataframe format for EVTVL")
+
+    return source_list
